# src/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "TeLoCobro - OpenAPI 3.1.1",
    "description": "La API de TeLoCobro ofrece una interfaz robusta y flexible que permite a los desarrolladores interactuar con TeLoCobro de manera eficiente. Esta API está diseñada para facilitar el acceso y la gestión de datos, permitiendo realizar operaciones como consultar, actualizar, eliminar y crear información. Con un enfoque RESTful y una comunicación basada en HTTP(S), la API está optimizada para integrarse fácilmente en diversas plataformas y aplicaciones.\n\n**Developer:** <br> Name: Carlos Daniel Menchaca Arauz <br> Linkedin: [@iamcarlosdaniel](https://www.linkedin.com/in/iamcarlosdaniel/) <br> Portfolio: [iamcarlosdaniel.com](https://iamcarlosdaniel.com)\n\n**Testers:**\n\nName: name <br> Linkedin: [@link]()\n\nName: name <br> Linkedin: [@link]()\n\nDocumentation written by <br> **Carlos Daniel Menchaca Arauz**",
    "version": "1.0.11"
  },
...
  "servers": [
    {
      "url": "http://localhost:3000/api/v1"
    }
  ],
...
  "paths": {
...
    "/users/me": {
      "get": {
        "tags": ["Users"],
        "summary": "Get profile",
        "description": "Permite a un usuario obtener la informacion de su perfil.",
        "operationId": "getProfile",
        "security": [{ "auth_token": [] }, { "auth_cookie": [] }],
        "responses": {
          "200": {
            "description": "Perfil obtenido exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "description": "Estado de la respuesta",
                      "example": "OK"
                    },
                    "data": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        }
      },
      "put": {
        "tags": ["Users"],
        "summary": "Update profile",
        "description": "Permite a un usuario actualizar la informacion de su perfil.",
        "operationId": "updateProfile",
        "security": [{ "auth_token": [] }, { "auth_cookie": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "Nombre de usuario"
                  },
                  "first_name": {
                    "type": "string",
                    "description": "Nombres del usuario"
                  },
                  "last_name": {
                    "type": "string",
                    "description": "Apellidos del usuario"
                  },
                  "date_of_birth": {
                    "type": "string",
                    "description": "Fecha de nacimiento del usuario"
                  },
                  "gender": {
                    "type": "string",
                    "enum": ["male", "female", "other"],
                    "description": "Género del usuario"
                  },
                  "bio": {
                    "type": "string",
                    "description": "Biografía del usuario"
                  }
                },
                "required": ["username", "email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Perfil actualizado exitosamente",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/users/me/email": {
      "post": {
        "tags": ["Users"],
        "summary": "Change email",
        "description": "Permite al usuario actualmente autenticado solicitar cambiar su correo electrónico.",
        "operationId": "changeEmail",
        "security": [{ "auth_token": [] }, { "auth_cookie": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "new_email": {
                    "type": "string",
                    "format": "email",
                    "description": "Nuevo correo electrónico del usuario"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "description": "Contraseña del usuario para verificar la identidad"
                  }
                },
                "required": ["new_email", "password"]
              }
            }
          },
          "responses": {
            "200": {
              "$ref": "#/components/responses/InternalServerErrorResponse"
            }
          }
        }
      }
    },
    "/users/me/email/confirm": {
      "patch": {
        "tags": ["Users"],
        "summary": "Confirm email change",
        "description": "Permite al usuario actualmente autenticado confirmar el cambio de su correo electrónico.",
        "operationId": "confirmEmailChange",
        "security": [{ "auth_token": [] }, { "auth_cookie": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "otp": {
                    "type": "string",
                    "description": "OTP de verificación enviado al nuevo correo electrónico"
                  }
                },
                "required": ["otp"]
              }
            }
          },
          "responses": {
            "200": {
              "$ref": "#/components/responses/InternalServerErrorResponse"
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "auth_token_header": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "in": "header",
        "name": "Authorization"
      },
      "auth_token_cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "auth_token",
        "description": "El token en esta cookie es un JWT."
      }
    },
    "responses": {
      "InternalServerErrorResponse": {
        "description": "Error interno del servidor",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "FAILED"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string",
                            "example": "Ocurrió un error al procesar su solicitud. Por favor intente de nuevo más tarde."
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}